Refactor: cts: Move validation code into a new python module.
This is also going to be useful for running command line tool regression
tests, so split the code out into a separate function and make it more
generic.

* Raise FileNotFoundError if the validator is not found so we can tell
  that error apart from validation errors.

* Add the xmllint return code to XmlValidationError so we can access it
  later if needed.

* Return the xmllint output instead of printing it so the caller can
  decide what to do with it.

* Use configure.ac to find the path to xmllint.
